requirements-docker.txt
# Docker 环境依赖（排除 Windows 专属的 pywin32）
openai>=1.0.0
toml>=0.10.0
fastapi>=0.100.0
uvicorn>=0.20.0
numpy>=1.24.0
python-docx>=1.0.0
